Moved api libs to the api folder
