+++ src/pages/[lang]/tshirt.js
@@ -1,8 +1,8 @@
 import { getLocaleStaticPaths, getLocaleByPath } from '../../lib/api/locale'
-import getTranslationsByLocale from '../../lib/translations'
+import getTranslationsByLocale from '../../lib/api/translations'
+import graphqlFetch from '../../lib/api/graphql-fetch'
+import getEntitiesData from '../../lib/api/get-entities-data'
 import html from '../../translations/en-US/introductions/tshirt.md'
-import graphqlFetch from '../../lib/graphql-fetch'
-import getEntitiesData from '../../lib/get-entities-data'
 
 import Layout from 'core/Layout'
 import PageFooter from 'core/pages/PageFooter'
